be/esport/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-jazzmin==3.0.1
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-decouple==3.8
sqlparse==0.5.2
watchdog==6.0.0
